requirements.txt
gunicorn
django-heroku
asgiref==3.2.7
beautifulsoup4==4.8.2
cycler==0.10.0
dj-database-url==0.5.0
django-background-task==0.1.8
django-background-tasks==1.2.5
django-bootstrap-select==0.1.3
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-dynamic-model==0.1.0
imbalanced-learn==0.6.2
imblearn==0.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
psycopg2-binary==2.8.4
psycopg2==2.7.5
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.3.1
whitenoise==5.0.1
Django==3.0.4
